Consolidated testcases and updated related docs
