--- tests/transceiver/system/process_restart/test_pmon_restart.py
@@ -1,108 +1,0 @@
-"""System / Process Restart - pmon daemon restart validation.
-
-Implements the pmon restart test from
-    ``docs/testplan/transceiver/system_test_plan.md``.
-
-Execution order::
-
-  session start
-    `- check_links_up()             <- session-scoped via
-                                        ``links_verified`` in
-                                        tests/transceiver/conftest.py
-                                        (failure skips every System test)
-    `- test_system_pmon_restart
-         |- <body>: restart pmon -> verify all ports recovery
-  session end
-    `- _system_post_session_checks (system/conftest.py)
-         |- post_state_restoration()
-         `- final link + LLDP check
-
-Failure handling: failures are accumulated per test case and reported in a
-single ``pytest.fail`` at the end, so a single run surfaces all issues across
-all ports.
-"""
-import logging
-import time
-import pytest
-
-from tests.transceiver.attribute_parser.attribute_keys import (
-    SYSTEM_ATTRIBUTES_KEY
-)
-from tests.transceiver.common.health_checks import capture_baseline
-from tests.transceiver.common.prerequisites import (
-    check_links_up
-)
-from tests.transceiver.common.verification import (
-    standard_port_recovery_and_verification
-)
-from tests.common.helpers.sonic_db import AppDbCli as sdbHelp
-
-logger = logging.getLogger(__name__)
-
-
-@pytest.mark.disable_loganalyzer
-def test_system_pmon_restart(
-    duthost, port_attributes_dict, expected_pid_changes,
-    lport_to_first_subport_mapping,
-):
-    """Restart pmon and verify all ports recover cleanly.
-
-    Implements the test described in
-    ``docs/testplan/transceiver/system_test_plan.md``.
-
-    Simple pmon restart:
-        1. Verify all ports are operationally up and record link up time
-        2. Restart pmon using 'sudo systemctl restart pmon'
-        3. Wait for pmon_restart_settle_sec
-        4. Execute Standard Port Recovery and Verification Procedure
-           for all ports
-        5. Verify pmon has been running for at least pmon_restart_settle_sec
-    """
-    expected_pid_changes.add("xcvrd")
-    ports = sorted(port_attributes_dict.keys())
-    assert ports, "port_attributes_dict is empty - nothing to validate"
-    health_baseline = capture_baseline(duthost)
-    failures = []
-
-    logger.info("Recording link states and uptime for %d port(s)", len(ports))
-    link_check = check_links_up(duthost, port_attributes_dict)
-    if not link_check["passed"]:
-        logger.warning("Validation on Start FAILED: some ports are down")
-    else:
-        appl_db = sdbHelp(duthost)
-        for port in ports:
-            logger.info(
-                "Recording initial link uptime: %s",
-                appl_db.hget_key_value(
-                    "PORT_TABLE:{}".format(port), "last_up_time"
-                ),
-            )
-
-    logger.info("Restarting pmon...")
-    duthost.restart_service('pmon')
-    pmon_wait = port_attributes_dict[ports[0]].get(
-        SYSTEM_ATTRIBUTES_KEY, {}
-    ).get("pmon_restart_settle_sec", 120)
-
-    # accounts for minimum timeout behavior of SPRaV
-    time.sleep(pmon_wait + 60)
-
-    # Wait for settle time and verify
-    result = standard_port_recovery_and_verification(
-        duthost, ports, port_attributes_dict,
-        link_up_timeout_sec=pmon_wait,
-        health_baseline=health_baseline,
-        lport_to_first_subport_mapping=lport_to_first_subport_mapping,
-        expected_pid_changes=expected_pid_changes,
-    )
-    if not result["passed"]:
-        failures.append(f"[startup] {result['details']}")
-        logger.warning("Startup validation FAILED: %s", result["details"])
-    else:
-        logger.info("Startup validation PASSED: %s", result["details"])
-
-    if failures:
-        pytest.fail(
-            f"pmon restart recovery FAILED on {len(failures)} port(s): \n - "
-            + "\n  - ".join(failures)
-        )
