--- tests/transceiver/system/process_restart/test_swss_restart.py
@@ -1,137 +1,0 @@
-"""System / Process Restart - swss daemon restart validation.
-
-Implements the swss restart test from
-    ``docs/testplan/transceiver/system_test_plan.md``.
-
-Execution order::
-
-  session start
-    `- check_links_up()             <- session-scoped via
-                                        ``links_verified`` in
-                                        tests/transceiver/conftest.py
-                                        (failure skips every System test)
-    `- test_system_swss_restart
-         |- <body>: restart swss -> verify pmon restart (if expected)
-         `-                      -> verify all ports recovery
-  session end
-    `- _system_post_session_checks (system/conftest.py)
-         |- post_state_restoration()
-         `- final link + LLDP check
-
-Failure handling: failures are accumulated per test case and reported in a
-single ``pytest.fail`` at the end, so a single run surfaces all issues across
-all ports.
-"""
-import logging
-import time
-import pytest
-
-from tests.transceiver.attribute_parser.attribute_keys import (
-    SYSTEM_ATTRIBUTES_KEY
-)
-from tests.transceiver.common.health_checks import capture_baseline
-from tests.transceiver.common.prerequisites import (
-    check_links_up
-)
-from tests.transceiver.common.verification import (
-    standard_port_recovery_and_verification
-)
-from tests.common.helpers.sonic_db import AppDbCli as sdbHelp
-from tests.common.platform.processes_utils import (
-    check_docker_uptime_minutes,
-    check_pmon_uptime_minutes
-)
-
-logger = logging.getLogger(__name__)
-
-
-@pytest.mark.disable_loganalyzer
-def test_system_swss_restart(
-    duthost, port_attributes_dict, expected_pid_changes,
-    lport_to_first_subport_mapping,
-):
-    """Restart swss and verify all ports recover cleanly.
-
-    See the module docstring for the full execution tree.  Steps:
-
-      * verify all ports are oper-up and record link-up timestamps,
-      * restart swss via ``sudo systemctl restart swss``,
-      * if ``expect_pmon_restart_with_swss_or_syncd`` is set, verify pmon
-        restarted as expected,
-      * wait for ``swss_restart_settle_sec`` then run Standard Port Recovery
-        and Verification for every port.
-
-    All (port, step) failures are accumulated and reported in a single
-    ``pytest.fail`` so one run surfaces every issue.
-    """
-    expected_pid_changes.add("xcvrd")
-    expected_pid_changes.add("swss")
-    expected_pid_changes.add("syncd")
-    expected_pid_changes.add("orchagent")
-    ports = sorted(port_attributes_dict.keys())
-    assert ports, "port_attributes_dict is empty - nothing to validate"
-    health_baseline = capture_baseline(duthost)
-    failures = []
-
-    logger.info("Recording link states and uptime for %d port(s)", len(ports))
-    link_check = check_links_up(duthost, port_attributes_dict)
-    if not link_check["passed"]:
-        logger.warning("Validation on Start FAILED: some ports are down")
-    else:
-        appl_db = sdbHelp(duthost)
-        for port in ports:
-            logger.info(
-                "Recording initial link uptime: %s",
-                appl_db.hget_key_value(
-                    "PORT_TABLE:{}".format(port), "last_up_time"
-                ),
-            )
-
-    logger.info("Restarting swss...")
-    duthost.restart_service('swss')
-
-    if check_docker_uptime_minutes(duthost, "swss", minimal_runtime=3):
-        failures.append("swss did not restart as expected")
-        logger.warning("swss FAILED to restart")
-
-    # Check whether pmon restarted alongside swss
-    if port_attributes_dict[ports[0]].get(
-        SYSTEM_ATTRIBUTES_KEY, {}
-    ).get("expect_pmon_restart_with_swss_or_syncd", False):
-        time.sleep(15)
-        logger.info("Verifying pmon restart after swss restart...")
-        if check_pmon_uptime_minutes(duthost, minimal_runtime=3):
-            failures.append("[pmon] pmon did not restart as expected")
-            logger.warning(
-                "pmon FAILED to Restart when"
-                " expect_pmon_restart_with_swss_or_syncd is True"
-            )
-
-    syncd_wait = port_attributes_dict[ports[0]].get(
-        SYSTEM_ATTRIBUTES_KEY, {}
-    ).get("syncd_restart_settle_sec", 240)
-    swss_wait = port_attributes_dict[ports[0]].get(
-        SYSTEM_ATTRIBUTES_KEY, {}
-    ).get("swss_restart_settle_sec", 180)
-    max_wait = max(syncd_wait, swss_wait)
-    time.sleep(max_wait)
-
-    # Wait for settle time and verify
-    result = standard_port_recovery_and_verification(
-        duthost, ports, port_attributes_dict,
-        link_up_timeout_sec=max_wait,
-        health_baseline=health_baseline,
-        lport_to_first_subport_mapping=lport_to_first_subport_mapping,
-        expected_pid_changes=expected_pid_changes,
-    )
-    if not result["passed"]:
-        failures.append(f"[startup] {result['details']}")
-        logger.warning("Startup validation FAILED: %s", result["details"])
-    else:
-        logger.info("Startup validation PASSED: %s", result["details"])
-
-    if failures:
-        pytest.fail(
-            f"swss restart recovery FAILED on {len(failures)} port(s):\n  - "
-            + "\n  - ".join(failures)
-        )
